a lot of changes)

+++ src/main/java/com/mmdev/dictionaryy/repository/AdminRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Integer> {}
+public interface AdminRepository extends JpaRepository<Admin, Long> {}
